+++ features/support/env.rb
@@ -10,7 +10,7 @@ DB_NAME           = ENV.fetch('TEST_DB_NAME')
 DB_USER           = ENV.fetch('TEST_DB_USER')
 DB_PASS           = ENV.fetch('TEST_DB_PASSWORD')
 DB_PORT           = ENV.fetch('TEST_DB_PORT')
-DB_HOST           = ENV.fetch('TEST_DB_HOSTNAME', 'localhost')
+DB_HOST           = ENV.fetch('TEST_DB_HOSTNAME', '127.0.0.1')
 CONNECTION_STRING = "#{DB_USER}:#{DB_PASS}@tcp(#{DB_HOST}:#{DB_PORT})/#{DB_NAME}?tls=skip-verify"
 TMP_DIR           = '/tmp/'
 FILE_NAME         = 'ruby_encrypted'
@@ -21,10 +21,11 @@ Before do |_scenario|
     config.service_name = SERVICE_NAME
     config.product_id = PRODUCT_ID
     config.metastore = METASTORE
+    config.sql_metastore_db_type = 'mysql'
     config.connection_string = CONNECTION_STRING
     config.kms = KMS
     config.enable_session_caching = true
-    config.verbose = false
+    config.verbose = true
   end
 end